florida-game
In this game, there will be a bucket, crabs and seashells in a
beach setting. The player will have to collect the seashells using
the bucket. If they miss, they loose one point, if they collect a
seashell, they gain one point. If you touch the crab with the bucket,
you loose. You need 20 points to win the game.